+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@kafkajs/confluent-schema-registry": "^3.6.1",
     "exponential-backoff": "^3.1.2",
+    "jsdoc": "^4.0.4",
     "node-rdkafka": "^3.2.1"
   },
   "devDependencies": {
